+++ src/api/products/products.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common"
-import { AirlakeModule } from "./airlake/airlake.module"
+import { DatalakeModule } from "./datalake/datalake.module"
 import { FrostlakeModule } from "./frostlake/frostlake.module"
 import { SwapModule } from "./swap/swap.module"
 import { PayModule } from "./pay/pay.module"
@@ -8,7 +8,7 @@ import { HyperedgeModule } from "./hyperedge/hyperedge.module"
 import { LedgerscanModule } from "./ledgerscan/ledgerscan.module"
 
 @Module({
-  imports: [AirlakeModule, FrostlakeModule, SwapModule, PayModule, NftstudioModule, HyperedgeModule, LedgerscanModule],
+  imports: [DatalakeModule, FrostlakeModule, SwapModule, PayModule, NftstudioModule, HyperedgeModule, LedgerscanModule],
 })
 
 export class ProductsModule { }