+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Rebranding of Swapstream as Lenstack Swap
 * Rebranding of Edgescan as Lenstack Ledgerscan
 * Rebranding of Snowlake as Lenstack NFT Studio
+* Rebranding of Airlake as Lenstack Datalake
 
 ### v3.0
 * Created DTO for all controllers wherever required to avoid raw data extraction from request body